pyproject.toml
[project]
name = "clear"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jayden Pyles", email = "jpylesbusiness@gmail.com"},
]
dependencies = [
    "boto3>=1.35.34",
    "pyaml>=24.9.0",
    "boto3-stubs>=1.35.34",
    "b2sdk>=2.5.1",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
